Refactor top-level import structure
