+++ examples/slider-example.py
@@ -5,8 +5,8 @@ are connected to the same value.
 from observ import reactive
 from PySide6 import QtCore, QtWidgets
 
-from collagraph import Collagraph, create_element as h, EventLoopType
-from collagraph.renderers import PySideRenderer
+import collagraph as cg
+from collagraph import h
 
 
 def Example(props):
@@ -66,7 +66,9 @@ def Example(props):
 if __name__ == "__main__":
     app = QtWidgets.QApplication()
 
-    gui = Collagraph(renderer=PySideRenderer(), event_loop_type=EventLoopType.QT)
+    gui = cg.Collagraph(
+        renderer=cg.PySideRenderer(), event_loop_type=cg.EventLoopType.QT
+    )
     state = reactive({"value": 50, "tracking": True})
     element = h(Example, state)
